+++ base/src/main/java/proguard/classfile/JavaVersionConstants.java
@@ -53,4 +53,5 @@ public class JavaVersionConstants {
   public static final String CLASS_VERSION_21 = "21";
   public static final String CLASS_VERSION_22 = "22";
   public static final String CLASS_VERSION_23 = "23";
+  public static final String CLASS_VERSION_24 = "24";
 }